Only display entries and notes to owner in index
Implement ownership in Entry model, and query the model filtered by
current user id in the view. This ensures that the entries passed to the
template are owned by the current user.
The query to the Notes model is filtered by the ids of the entries
belonging to the current user.