Fix vunerability to change FK of notes
When manually entering the id of an entry in the URL it was possible to
change the foreign key of a note.

Create a function to check if an entry is the FK value of a note, and
call it in the view after asserting the current user is the owner of the
entry.